refactor(browser): split BrowserStore into per-domain mixins (B4)
